==> 5. Holt-Winters' forecast, using frequencies obtained from R analysis.ipynb <==
### Summary

Results submitted to the contest page on 29 Jan 2020 18:50:18, achieving a $100 \times (1-SMAPE)$ public score of 70.04

import os
import numpy as np
import pandas as pd
import seaborn as sns
import datetime as dt
import matplotlib.pyplot as plt
from statsmodels.tsa.seasonal import seasonal_decompose
from statsmodels.tsa.holtwinters import ExponentialSmoothing

#sns.set_style('whitegrid')
#sns.set(rc={'figure.figsize':(16, 4)})
plt.rc('figure',figsize=(16,4))
plt.rc('font',size=13)

# main dataset to calculate offsets
train = pd.read_pickle('train.pkl')
test = pd.read_pickle('test.pkl')

trainx = pd.read_csv('train_1.csv', index_col=0)
test = pd.read_csv('test_forR.csv', index_col=0)

train_orig = train.copy()

trainx_orig = trainx.copy()
test_orig = test.copy()

# to reset 
trainx = trainx_orig.copy()
test = test_orig.copy()

sat_ids = test['sat_id'].unique()

# number of observations in test for each satellite n
num_test = [(test['sat_id'] == n).sum() for n in sat_ids]

# dataframe of frequencies for each satellite, found by R's 'findfrequency'
freq = pd.read_csv('frequencies.csv').drop('Unnamed: 0', axis=1)

freq.head()

freq.columns

freq.shape

trainx.shape, test.shape

trainx.head()

test.head()

# reindex to epoch 
trainx.index = trainx['epoch']
test.index = test['epoch']

trainx.head()

test.head()

train.head()

def calculate_offset(k):
    
    offsets = []
    for n in sat_ids:
        satdata_n = train[train['sat_id'] == n][k]
        off_n = max(satdata_n) - min(satdata_n)
        offsets.append(off_n)
    
    return offsets

# the offsets for x are stored in off_x
off = calculate_offset(k='x')

off[0]

**Create a function that takes in x values from january and forecasts for the next month for all 300 satellites,
using the frequencies specified by the freq dataframe**


- grab data
- fit model to that data
- forecast for that data
- deoffset it
- put it into a list
- concatenate all forecast data into a single list and attach it to Test

# ETS(M,A_d,M) is multiplicative error, damped additive trend, multiplicative seasonality
# this makes forecast for all 300 satellites, use forecast_plot to make forecast for one satellite

def forecast(k):
    
    forecast_data = list()
    
    for i in range(0,len(sat_ids)):
        
        n = sat_ids[i]                # n is the satellite ID number
        frequency_n = freq['f_'+k][i] # time series frequency for Satellite n
        offset_n = off_k[i]             # offset for Satellite n
        num_test_n = num_test[i]      # number of observations that will be forecasted for Satellite n
    
        satdata_n = train_k[train_k['sat_id'] == n][k] # data for kinematic state k, Satellite n
    
        fit_n = ExponentialSmoothing(np.asarray(satdata_n), trend='add', seasonal='mul',
                                     seasonal_periods = frequency_n, damped=True).fit()
        
        forecast_n = fit_n.forecast(num_test_n)
                               
        forecast_n -= offset_n
        
        forecast_data.append(forecast_n)
        
        print(f'Forecast of {k} for Satellite {n} has been done.')
    
    return forecast_data

Cell below is a small example to understand how to loop over the satellites. 

# For each satellite number n from the 300 unique satellite IDs in the Test dataset,
#forecast_x = list()

#for i in range(0,2):
    
#    n = sat_ids[i]
    
#    frequency_n = freq['f_'+'x'][i]
#    offset_n = off[i]
#    num_test_n = num_test[i]
    
    # grab nth satellite data
#    satdata_n = trainx[trainx['sat_id'] == n].loc[:,'x']
    
    # fit model to that data
#    fit_n = ExponentialSmoothing(np.asarray(satdata_n), trend = 'add', seasonal = 'mul',
#                                 seasonal_periods = frequency_n, damped = True).fit()
    
    # use model to produce forecast data for the epochs in Test
#    forecast_n = fit_n.forecast(num_test_n)
    
    # deoffset the forecast data by the corresponding amount in 'off'
#    forecast_n -= offset_n
    
    # put the forecast data into a list
#    forecast_x.append(forecast_n)
    

num_test[0:2]

#%%time
#forecast_x = forecast(k='x')

%%time
forecast_x = forecast(k='x')

# forecast for remaining kinematic states
rem = 'y z Vx Vy Vz'.split()

# after correcting for seasonality = 1 error, start forecasting from Vx
rem = 'Vx Vy Vz'.split()

forecast_list = list()

%%time
forecast_list = list()

for i in range(0,3):
    
    kinem = rem[i]
    
    train_k = pd.read_csv(f'train_{i+4}.csv', index_col=2)
    
    off_k = calculate_offset(k=kinem)
    
    forecast_k = forecast(k=kinem)
    
    forecast_list.append(forecast_k)

- Seasonal periods must be larger than 1. Forecasts for x, y, z went OK but some seasonal periods for Vx, Vy, Vz must be greater than 1 for the forecast to be done.
- Python raised quite a few ConvergenceWarnings because the optimizer failed to find a local/global minimum. Is this because of too few data points for some of the satellites?


# find which variables has frequency 1
(freq == 1).sum()

# change them to 2
for k in freq.columns[3:]:
    
    # DO THIS
    freq.loc[freq[k] == 1, k] = 2
    
    # DONT DO THIS
    #freq[freq[k] == 1][k] = 2
    

freq.head()

# this forecast_list was after forecasting for y and z
forecast_y = forecast_list[0]

forecast_z = forecast_list[1]

# this forecast_list was after correcting for seasonality = 1 and forecasting for Vx, Vy, Vz
forecast_Vx = forecast_list[0]
forecast_Vy = forecast_list[1]
forecast_Vz = forecast_list[2]

len(forecast_list)

len(forecast_x)

# concatenate all lists inside each forecast_k
import itertools

# hats on the forecasted data
x_hat = list(itertools.chain.from_iterable(forecast_x))
y_hat = list(itertools.chain.from_iterable(forecast_y))
z_hat = list(itertools.chain.from_iterable(forecast_z))

Vx_hat = list(itertools.chain.from_iterable(forecast_Vx))
Vy_hat = list(itertools.chain.from_iterable(forecast_Vy))
Vz_hat = list(itertools.chain.from_iterable(forecast_Vz))

test.index = test['epoch']
train.index = train['epoch']

test['x'] = x_hat
test['y'] = y_hat
test['z'] = z_hat
test['Vx'] = Vx_hat
test['Vy'] = Vy_hat
test['Vz'] = Vz_hat

# make a copy for submitting results
test_sub = test.copy()

test_sub.head()

submission = test_sub[test.columns[3:]]

submission.head()

submission.to_csv('submission.csv')

This was submitted to the contest page on 29 Jan 2020 18:50:18, achieving a $100 \times (1-SMAPE)$ public score of 70.04
